corregido

+++ G26-HELOU-MAO-TP_FINAL/G26-HELOU-MAO-TP_FINAL/cMedico.cpp
@@ -1,11 +1,12 @@
 #include "cMedico.h"
 
-cMedico::cMedico(string miNombre, string miApellido, float miNumero)
+cMedico::cMedico(string miNombre, string miApellido, int miNumero)
 {
 	this->aNombre = miNombre;
 	this->aApellido = miApellido;
 	this->aNroEmplpeado = miNumero;
 }
+
 cMedico::~cMedico() {
 
 }
